requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-experimental
faiss-cpu
rank_bm25
flashrank
sentence-transformers
fastapi
uvicorn
streamlit
pypdf
fpdf2
pandas
numpy
python-dotenv
requests
beautifulsoup4
